==> EPIC024/notebooks/2.0-mtm-preprocessing-ppnn.ipynb <==
# 1. Librerias

import pandas as pd
import numpy as np
import seaborn as sns

import matplotlib.pyplot as plt
from scipy import stats

import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

# 2. Data 

dat=pd.read_csv('../data/modeling/01_raw.csv',sep='|',encoding='latin-1')
dat.head()

nat=dat[dat.TIPPER=='P']
nat.shape

# 3. Limpieza de datos

nat.isna().sum()/len(nat)

# 4. Universo + Split (Train - Test)

nat.to_csv(r"../data/modeling/02_preprocessed_ppnn.csv",index=False,sep="|")

# Train
train = nat[nat.PERIODO!=202310]
train.shape

train.to_csv(r"../data/modeling/03_train_ppnn.csv",index=False,sep="|")

# Test
test = nat[(nat["PERIODO"]==202310)]
test.shape

test.to_csv(r"../data/modeling/03_test_ppnn.csv",index=False,sep="|")
