==> EPIC024/notebooks/3.0-mtm-modeling-ppjj.ipynb <==
# 1. Librerías

import pandas as pd
import numpy as np

from sklearn.preprocessing import scale
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import KMeans
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import silhouette_score
from sklearn.ensemble import IsolationForest
from sklearn import metrics
import shap
#from utils.Validator import *
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)
pd.set_option('display.float_format', lambda x: '%.4f' % x)

# 2. Data

dat= pd.read_csv(r"../data/modeling/03_train_ppjj.csv",sep="|")
dat.head()

# 3. Escalón 2 - Población con riesgo

esc2=dat[(dat.MTO_INGRESOS_MES>=1100)&(dat.MTO_INGRESOS_MES<=30e+6)]

# 4. Escalón 3 - Directo a evaluación

esc3=dat[dat.MTO_INGRESOS_MES>30e+6]

# 5. Modelamiento

## Modelo 1: k-MEANS

#### Con las variables originales

x_model=['FLG_PERFIL_INGRESOS_3DS', 'MTO_CONOTROSPROXIMOS', 'CTD_CONOTROSPROXIMOS']

train_jur_std = StandardScaler().fit_transform(esc2[x_model])
k = [2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20]
for ki in k:
        outputKMeans = KMeans(n_clusters = ki, max_iter=300,init='k-means++', random_state = 7).fit(train_jur_std)
        scoreSilhoutte = metrics.silhouette_score(train_jur_std, outputKMeans.labels_, metric='euclidean')  
        print(" k=", ki," SC=", scoreSilhoutte)

#### Reemplazando var mto_conotrosproximos con mto_ingresos_mes

x_model_2=['FLG_PERFIL_INGRESOS_3DS','CTD_CONOTROSPROXIMOS','MTO_INGRESOS_MES']

train_jur_std2 = StandardScaler().fit_transform(esc2[x_model_2])
k = [2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20]
for ki in k:
        outputKMeans = KMeans(n_clusters = ki, max_iter=300,init='k-means++', random_state = 7).fit(train_jur_std2)
        scoreSilhoutte = metrics.silhouette_score(train_jur_std2, outputKMeans.labels_, metric='euclidean')  
        print(" k=", ki," SC=", scoreSilhoutte)

## Modelo 2: Isolation Forest

#### Con las variables originales

# Evaluación del mejor número de árboles

from matplotlib.pyplot import step
silhouette_list = []

for est in np.arange(100,800,100):
    cluster_labels = IsolationForest(n_estimators = est,
                                    n_jobs=-1,
                                    max_features=(len(esc2[x_model].columns)),
                                    contamination=0.01,
                                    random_state=66,bootstrap=True).fit_predict(esc2[x_model])
    out_n = cluster_labels[cluster_labels==-1].shape[0]
    sil = silhouette_score(esc2[x_model],cluster_labels, metric='euclidean')
    silhouette_list.append(sil)
    print("silhouette for "+str(est)+" with n=" +str(out_n)+" outliers is: "+str(sil))

from matplotlib.pyplot import step
silhouette_list = []

for cont in np.arange(0.001,0.011,0.001):
    cluster_labels = IsolationForest(n_estimators = 400,
                                    n_jobs=-1,
                                    max_features=(len(esc2[x_model].columns)),
                                    contamination=cont,
                                    random_state=66,bootstrap=True).fit_predict(esc2[x_model])
    out_n = cluster_labels[cluster_labels==-1].shape[0]
    sil = silhouette_score(esc2[x_model],cluster_labels, metric='euclidean')
    silhouette_list.append(sil)
    print("silhouette for "+str(np.round(cont,4))+" with n=" +str(out_n)+" outliers is: "+str(sil))

#### Reemplazando var mto_conotrosproximos con mto_ingresos_mes

x_model_2=['FLG_PERFIL_INGRESOS_3DS',
 'CTD_CONOTROSPROXIMOS',
 'MTO_INGRESOS_MES']

from matplotlib.pyplot import step
silhouette_list = []

for cont in np.arange(0.001,0.006,0.001):
    cluster_labels = IsolationForest(n_estimators = 400,
                                    n_jobs=-1,
                                    max_features=(len(esc2[x_model_2].columns)),
                                    contamination=cont,
                                    random_state=66,bootstrap=True).fit_predict(esc2[x_model_2])
    out_n = cluster_labels[cluster_labels==-1].shape[0]
    sil = silhouette_score(esc2[x_model_2],cluster_labels, metric='euclidean')
    silhouette_list.append(sil)
    print("silhouette for "+str(np.round(cont,4))+" with n=" +str(out_n)+" outliers is: "+str(sil))

#### Normalizando

from matplotlib.pyplot import step

df2 = esc2[x_model]

scaler = MinMaxScaler()
df2[x_model] = scaler.fit_transform(df2)
df2=df2[x_model]

silhouette_list = []

for cont in np.arange(0.01,0.15,0.01):
    cluster_labels = IsolationForest(n_estimators = 500,
                                    n_jobs=-1,
                                    max_features=(len(df2.columns)),
                                    contamination=cont,
                                    random_state=66).fit_predict(df2)
    out_n = cluster_labels[cluster_labels==-1].shape[0]
    sil = silhouette_score(df2,cluster_labels, metric='euclidean')
    silhouette_list.append(sil)
    print("silhouette for "+str(np.round(cont,4))+" with n=" +str(out_n)+" outliers is: "+str(sil))

#### Estandarizando

df2 = esc2[x_model]

df2[x_model] = StandardScaler().fit_transform(df2[x_model])
df2=df2[x_model]

silhouette_list = []

for cont in np.arange(0.01,0.1,0.01):
    cluster_labels = IsolationForest(n_estimators = 500,
                                    n_jobs=-1,
                                    max_features=(len(df2.columns)),
                                    contamination=cont,
                                    random_state=66).fit_predict(df2)
    out_n = cluster_labels[cluster_labels==-1].shape[0]
    sil = silhouette_score(df2,cluster_labels, metric='euclidean')
    silhouette_list.append(sil)
    print("silhouette for "+str(np.round(cont,4))+" with n=" +str(out_n)+" outliers is: "+str(sil))

# Modelo Final

X_train = esc2[x_model]

iforest = IsolationForest(n_estimators = 400,
                                    n_jobs=-1,
                                    max_features=(len(X_train[x_model].columns)),
                                    contamination=0.002,
                                    random_state=66,bootstrap=True).fit(X_train[x_model])

esc2['label_iforest']=iforest.predict(X_train[x_model])

esc2[esc2.label_iforest==-1].PERIODO.value_counts()

esc2[esc2.label_iforest==-1].describe().T

#Guardamos el modelo
import pickle
pickle.dump(iforest, open(r"../src/02_models/iforest_model_ppjj", 'wb'))
