==> EPIC024/notebooks/3.0-mtm-modeling-ppnn.ipynb <==
# 1. Librerías

import pandas as pd
import numpy as np

from sklearn.preprocessing import scale
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import KMeans
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import silhouette_score
from sklearn.ensemble import IsolationForest
from sklearn import metrics
import shap
#from utils.Validator import *
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)
pd.set_option('display.float_format', lambda x: '%.4f' % x)

# 2. Data

dat= pd.read_csv(r"../data/modeling/03_train_ppnn.csv",sep="|")
dat.head()

# 3. Escalón 2 - Población con riesgo

esc2=dat[(dat.MTO_INGRESOS_MES>=150)&(dat.MTO_INGRESOS_MES<=1e+6)]

# 4. Escalón 3 - Directo a evaluación

esc3=dat[dat.MTO_INGRESOS_MES>1e+6]

# 5. Modelamiento

## Modelo 1: k-MEANS

#### Con las variables originales

x_model=['MTO_CONOTROSPROXIMOS', 'CTD_CONOTROSPROXIMOS']

# Se tomará una muestra del 30% ppor periodo
from sklearn.model_selection import train_test_split
muestra1, muestra2 = train_test_split(esc2, train_size=0.7, stratify=esc2['PERIODO'])

train_nat_std = StandardScaler().fit_transform(muestra2[x_model])

k = [3,4,5,6,7,8,9,10,11,12,13,14,15]
for ki in k:
        outputKMeans = KMeans(n_clusters = ki, init='k-means++', random_state = 101).fit(train_nat_std)
        scoreSilhoutte = metrics.silhouette_score(train_nat_std, outputKMeans.labels_, metric='euclidean')  
        print(" k=", ki," SC=", scoreSilhoutte)

#### Agregando FLG_PERFIL_INGRESOS_3DS

x_model2=['MTO_CONOTROSPROXIMOS', 'CTD_CONOTROSPROXIMOS','FLG_PERFIL_INGRESOS_3DS']

train_nat_std = StandardScaler().fit_transform(muestra2[x_model2])

k = [3,4,5,6,7,8,9,10]
for ki in k:
        outputKMeans = KMeans(n_clusters = ki, init='k-means++', random_state = 101).fit(train_nat_std)
        scoreSilhoutte = metrics.silhouette_score(train_nat_std, outputKMeans.labels_, metric='euclidean')  
        print(" k=", ki," SC=", scoreSilhoutte)

## Modelo 2: Isolation Forest

#### Con las variables originales

silhouette_list = []

for cont in np.arange(0.0001,0.0005,0.0001):
    cluster_labels = IsolationForest(n_estimators = 200,
                                    n_jobs=-1,
                                    max_features=(len(muestra2[x_model].columns)),
                                    contamination=cont,#proporcion de outliers en el dataset
                                    random_state=66,bootstrap=True).fit_predict(muestra2[x_model])
    
    out_n = cluster_labels[cluster_labels==-1].shape[0]
    sil = silhouette_score(muestra2[x_model],cluster_labels, metric='euclidean')
    silhouette_list.append(sil)
    print("silhouette for "+str(np.round(cont,4))+" with n=" +str(out_n)+" outliers is: "+str(sil))

#### Agregando FLG_PERFIL_INGRESOS_3DS

x_model2=['MTO_CONOTROSPROXIMOS', 'CTD_CONOTROSPROXIMOS','FLG_PERFIL_INGRESOS_3DS']

silhouette_list = []

for cont in np.arange(0.0001,0.0006,0.0001):
    cluster_labels = IsolationForest(n_estimators = 200,
                                    n_jobs=-1,
                                    max_features=(len(muestra2[x_model2].columns)),
                                    contamination=cont,#proporcion de outliers en el dataset
                                    random_state=66,bootstrap=True).fit_predict(muestra2[x_model2])
    
    out_n = cluster_labels[cluster_labels==-1].shape[0]
    sil = silhouette_score(muestra2[x_model2],cluster_labels, metric='euclidean')
    silhouette_list.append(sil)
    print("silhouette for "+str(np.round(cont,4))+" with n=" +str(out_n)+" outliers is: "+str(sil))

# Modelo Final

X_train = esc2[x_model2]

iforest = IsolationForest(n_estimators = 200,
                                    n_jobs=-1,
                                    max_features=(len(X_train[x_model2].columns)),
                                    contamination=0.0004,
                                    random_state=66,bootstrap=True).fit(X_train[x_model2])

esc2['if_ros']=iforest.predict(esc2[x_model2])

esc2[esc2.if_ros==-1].PERIODO.value_counts().mean()

esc2[esc2.if_ros==-1].describe().T

#Guardamos el modelo
import pickle
pickle.dump(iforest, open(r"../src/02_models/iforest_model_ppnn", 'wb'))
