==> EPIC024/notebooks/4.0-mtm-model-final-ppnn.ipynb <==
# 1. Librería

import pandas as pd
import numpy as np
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import silhouette_score
from sklearn.ensemble import IsolationForest
#from utils.Validator import *
import pickle
import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns', None)
pd.set_option('display.float_format', lambda x: '%.4f' % x)

# 2. Importación de la data

train = pd.read_csv(r"../data/modeling/03_train_ppnn.csv",sep="|")   
test = pd.read_csv(r"../data/modeling/03_test_ppnn.csv",sep="|") 

## Train

train_sc2=train[(train.MTO_INGRESOS_MES>=150)&(train.MTO_INGRESOS_MES<=1e+6)]

train_sc3 = train[train['MTO_INGRESOS_MES']>1e+6]
train_sc3.shape

train_sc3['n_cluster']=1

## Test

test_sc2=test[(test.MTO_INGRESOS_MES>=150)&(test.MTO_INGRESOS_MES<=1e+6)]

test_sc2.shape

test_sc3 = test[test['MTO_INGRESOS_MES']>1e+6]
test_sc3.shape

test_sc3['n_cluster']=1

# 2. Exportación del modelo

model = pickle.load(open(r"../src/02_models/iforest_model_ppnn", 'rb'))

model.feature_names_in_

# 3. Exportación de data

x_model=['MTO_CONOTROSPROXIMOS', 'CTD_CONOTROSPROXIMOS','FLG_PERFIL_INGRESOS_3DS']

vars_export=['PERIODO', 'CODCLAVECIC','TIPPER','MTO_INGRESOS_MES','FLG_PERFIL_INGRESOS_3DS','MTO_CONOTROSPROXIMOS', 'CTD_CONOTROSPROXIMOS','n_cluster']

# TRAIN
train_sc2['n_cluster']=model.predict(train_sc2[x_model])
#Alertas según el modelo Isolations 
alertas_train=train_sc2[train_sc2["n_cluster"]==-1]
alertas_train=pd.concat([alertas_train, train_sc3], ignore_index=True)
#Unimos las alertas del escalón 3 y 2
alertas_train[vars_export].to_csv(r"../data/modeling/04_predicted_train_ppnn.csv",index=False,sep="|")

# TEST
test_sc2['n_cluster']=model.predict(test_sc2[x_model])
#Alertas según el modelo Isolations 
alertas_test=test_sc2[test_sc2["n_cluster"]==-1]
alertas_test=pd.concat([alertas_test, test_sc3], ignore_index=True)
#Unimos las alertas del escalón 3 y 2
alertas_test[vars_export].to_csv(r"../data/modeling/04_predicted_test_ppnn.csv",index=False,sep="|")


alertas_test.shape # 16 modelo y 9 directo a evaluación

alertas_test.n_cluster.value_counts()
# -1 modelo
# 1 ouliers (DaE)
